Create a library to copy a CTF trace
Move all the code from the writer plugin to a standalone plugin library
so it can be used by other plugins.
