+++ plugins/Makefile.am
@@ -1,3 +1,3 @@
-SUBDIRS = ctf text muxer writer utils
+SUBDIRS = ctf text muxer utils libctfcopytrace writer
 
 noinst_HEADERS = plugins-common.h
