+++ plugins/writer/Makefile.am
@@ -1,4 +1,5 @@
-AM_CFLAGS = $(PACKAGE_CFLAGS) -I$(top_srcdir)/include -I$(top_srcdir)/plugins
+AM_CFLAGS = $(PACKAGE_CFLAGS) -I$(top_srcdir)/include -I$(top_srcdir)/plugins \
+	    -I$(top_srcdir)/plugins/libctfcopytrace
 
 SUBDIRS = .
 
@@ -16,4 +17,7 @@ libbabeltrace_plugin_ctf_writer_la_LDFLAGS = \
 
 libbabeltrace_plugin_ctf_writer_la_LIBADD = \
 	$(top_builddir)/lib/libbabeltrace.la \
-	$(top_builddir)/formats/ctf/libbabeltrace-ctf.la
+	$(top_builddir)/formats/ctf/libbabeltrace-ctf.la \
+	$(top_builddir)/plugins/libctfcopytrace/libctfcopytrace.la
+
+noinst_HEADERS = writer.h
